Add run configuration for running unit test with android studio instead of gradle.
By default if I run unit tests with that run button which is near them, those will run with gradle.
We should run tests with gradle when want to have a continues integration, but when we want to run tests locally we can
use another option that run tests faster. This option is using Junit run configuration.